craftsmen-marketplace-backend/pyproject.toml
[project]
name = "craftsmen-marketplace-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.2",
    "bcrypt>=4.3.0",
    "cryptography>=45.0.4",
    "facebook-sdk>=3.1.0",
    "fastapi>=0.115.14",
    "google-cloud-speech>=2.33.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "instagrapi>=2.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.3",
]
